refactor: extract findOrders function
