+++ packages/backend/src/orders-logic.js
@@ -1,0 +1,36 @@
+const QueryUtils = require('./query-utils')
+const { getMongoClient } = require('./utils')
+
+async function findOrders(filters) {
+  const client = await getMongoClient()
+  const queryFilters = []
+  if (filters.id)
+    queryFilters.push({
+      id: filters.id,
+    })
+  if (filters.phone) {
+    queryFilters.push({
+      $or: [
+        QueryUtils.includes('mainPhone', filters.phone),
+        QueryUtils.includes('additionalPhone', filters.phone),
+      ],
+    })
+  }
+  if (filters.city) queryFilters.push(QueryUtils.includes('city', filters.city))
+  if (filters.street)
+    queryFilters.push(QueryUtils.includes('street', filters.street))
+  if (filters.house)
+    queryFilters.push(QueryUtils.includes('house', filters.house))
+  return client
+    .db()
+    .collection('orders')
+    .find(
+      {
+        $and: queryFilters,
+      },
+      { limit: 50 }
+    )
+    .toArray()
+}
+
+module.exports = { findOrders }
